comments, added subscription function
